Getting started

Getting started

Installing

In Julia open the package manager in the REPL via ] and run:

(v1.0) pkg> add git@github.com:haampie/ArnoldiMethod.jl.git

Then use the package.

using ArnoldiMethod

Construct a partial Schur decomposition

ArnoldiMethod.jl exports the partialschur function which can be used to obtain a partial Schur decomposition of any matrix A.

ArnoldiMethod.partialschur — Function.

partialschur(A; nev, which, tol, mindim, maxdim, restarts) -> PartialSchur, History

Find nev approximate eigenpairs of A with eigenvalues near a specified target.

The matrix A can be any linear map that implements mul!(y, A, x), eltype and size.

The method will run iteratively until the eigenpairs are approximated to the prescribed tolerance or until restarts restarts have passed.

Arguments

The most important keyword arguments:

Keyword	Type	Default	Description

nev	Int	min(6, size(A, 1))	Number of eigenvalues

which	Target	LM()	One of LM(), LR(), SR(), LI(), SI(), see below.

tol	Real	√eps	Tolerance for convergence: ‖Ax - xλ‖₂ < tol * ‖λ‖

The target which can be any of subtypes(ArnoldiMethod.Target):

Target	Description

LM()	Largest magnitude: abs(λ) is largest

LR()	Largest real part: real(λ) is largest

SR()	Smallest real part: real(λ) is smallest

LI()	Largest imaginary part: imag(λ) is largest

SI()	Smallest imaginary part: imag(λ) is smallest

Note

The targets LI() and SI() only make sense in complex arithmetic. In real arithmetic λ is an eigenvalue iff conj(λ) is an eigenvalue and this conjugate pair converges simultaneously.

Return values

The function returns a tuple

decomp, history = partialschur(A, ...)

where decomp is a PartialSchur struct which forms a partial Schur decomposition of A to a prescribed tolerance:

> norm(A * decomp.Q - decomp.Q * decomp.R)

history is a History struct that holds some basic information about convergence of the method:

> history.converged
true
> @show history
Converged after 359 matrix-vector products

Advanced usage

Further there are advanced keyword arguments for tuning the algorithm:

Keyword	Type	Default	Description

mindim	Int	min(max(10, nev), size(A,1))	Minimum Krylov dimension (≥ nev)

maxdim	Int	min(max(20, 2nev), size(A,1))	Maximum Krylov dimension (≥ min)

restarts	Int	200	Maximum number of restarts

When the algorithm does not converge, one can increase restarts. When the algorithm converges too slowly, one can play with mindim and maxdim. It is suggested to keep mindim equal to or slightly larger than nev, and maxdim is usually about two times mindim.

source

From a Schur decomposition to an eigendecomposition

The eigenvalues and eigenvectors are obtained from the Schur form with the partialeigen function that is exported by ArnoldiMethod.jl:

λs, X = partialeigen(decomp::PartialSchur)

Note that whenever the matrix A is real-symmetric or Hermitian, the partial Schur decomposition coincides with the partial eigendecomposition, so in that case there is no need for the transformation.

Example

Here we compute the first ten eigenvalues and eigenvectors of a tridiagonal sparse matrix.

julia> using ArnoldiMethod, LinearAlgebra, SparseArrays
julia> A = spdiagm(
           -1 => fill(-1.0, 99),
            0 => fill(2.0, 100), 
            1 => fill(-1.0, 99)
       );
julia> decomp, history = partialschur(A, nev=10, tol=1e-6, which=SR());
julia> history
Converged after 178 matrix-vector products
julia> norm(A * decomp.Q - decomp.Q * decomp.R)
3.717314639756976e-8
julia> λs, X = partialeigen(decomp);
julia> norm(A * X - X * Diagonal(λs))
3.7173146389810755e-8